Spring 2 lesson02

+++ src/main/java/edu/geekbrains/dto/ProductDto.java
@@ -10,11 +10,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDto {
-
-
     private Long id;
     private String title;
-    private BigDecimal price;
-
-
+    private Integer price;
 }
